REST API - Enquiries & Model Update
Added:
REST API Controller for Enquiries:
  - Return list of EnquiryTypes available. (HTTPGET)

  - Add to EnquiryDB when a valid EnquiryDto is sent to the server (HTTPPOST)

- New Enquiry View:
  New Form for anonymous user to send their enquires which will then be mapped to an EnquiryDto and sent to the server DB via API POST

Changes:
Enquiry Model:
 - public byte ID --> public int ID

 - Code First Migration scripts has been altered.

+++ VideoRentalService/VideoRentalService/Models/Enquiry.cs
@@ -8,7 +8,7 @@ namespace VideoRentalService.Models
 {
     public class Enquiry
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
 
         [Required]
         [EmailAddress]
@@ -31,7 +31,6 @@ namespace VideoRentalService.Models
         public DateTime DateSubmitted { get; set; }
 
         public DateTime? DateResolved { get; set; }
-
 
         
     }
